backend/requirements.txt
# Core Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client Libraries
httpx>=0.25.2
aiohttp>=3.9.1

# External Service Integrations
PyGithub>=1.59.1
jira>=3.5.2
atlassian-python-api>=3.41.10
google-api-python-client>=2.110.0
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-generativeai>=0.3.2

# Python Requirements
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Date/Time
python-dateutil>=2.8.2

# Logging and Monitoring
structlog>=23.2.0